requirements.txt
anyio==4.7.0
certifi==2024.8.30
charset-normalizer==3.3.2
contourpy==1.3.0
cycler==0.12.1
fonttools==4.53.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.8
kiwisolver==1.4.5
matplotlib==3.9.2
numpy==2.1.0
packaging==24.1
pandas==2.2.2
pillow==10.4.0
PyJWT==2.9.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-telegram-bot==21.9
pytz==2024.1
pyupbit==0.2.34
requests==2.32.3
six==1.16.0
sniffio==1.3.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
websockets==13.0.1
